+++ scripts/kali-essentials.sh
@@ -1,0 +1,17 @@
+#Kali Essentials
+#This script does some `git-clone`ing to pull down tools I always carry with me on my Kali boxes:
+git clone https://github.com/lgandx/PCredz /opt/pcredz
+git clone https://bitbucket.org/LaNMaSteR53/recon-ng.git /opt/recon-ng
+git clone https://github.com/stufus/egresscheck-framework.git /opt/egress
+git clone https://github.com/drwetter/testssl.sh.git /opt/testssl
+git clone https://github.com/nccgroup/chuckle.git /opt/chuckle
+git clone https://github.com/stasinopoulos/commix.git /opt/commix
+git clone https://github.com/ChrisTruncer/EyeWitness /opt/eyewitness
+git clone https://github.com/SpiderLabs/Responder /opt/responder
+git clone https://github.com/1N3/Sn1per.git /opt/sniper
+git clone https://github.com/theZiz/aha.git /opt/aha
+git clone https://github.com/iSECPartners/sslyze.git /opt/sslyze
+git clone https://github.com/smicallef/spiderfoot.git /opt/spiderfoot
+git clone https://github.com/evilsocket/bettercap.git /opt/bettercap
+git clone https://github.com/byt3bl33d3r/MITMf.git /opt/mitmf
+git clone https://github.com/braimee/bpatty.git /opt/bpatty