+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Fixed link for [Robocopy](cmdline/windows/robocopy.md)
 * Added section on [Robocopy](cmdline/windows/robocopy.md)
 * Spruced up the [readme.md](README.md)
+* Clarification on which machine is PC1/PC2 in [egress-check](cmdline/linux/egress-check.md)
+* Addition of [kali-essentials](scripts/kali-essentials.sh) script